The sendString(), sendUint8(), sendUint32() functions return value 'int' changed to 'size_t'.

+++ utils.h
@@ -9,11 +9,11 @@
 #include <linux/in.h> //struct sockaddr_in
 
 bool setAddress(const std::string& address, struct sockaddr_in* local);
-int sendString(Socket& sock, const std::string& source);
+size_t sendString(Socket& sock, const std::string& source);
 int receiveString(Socket& sock, std::string& destination);
-int sendUint8(Socket& sock, uint8_t source);
+size_t sendUint8(Socket& sock, uint8_t source);
 int receiveUint8(Socket& sock, uint8_t& destination);
-int sendUint32(Socket& sock, uint32_t source);
+size_t sendUint32(Socket& sock, uint32_t source);
 int receiveUint32(Socket& sock, uint32_t& destination);
 
 enum Command {GET = 0, PUT = 1, LIST = 2, ERROR = 128, SUCCESS = 129, SEND_FILE = 130, SEND_LIST = 131};
